# RFC: Apple2: Handle IRQs without ProDOS