This makes IRQ handling very much faster on apple2enh.

**Tests:**
Tested with serial and mouse drivers:
- target `-t apple2` on emulated apple2e, apple2ee, apple2ep, apple2c0, apple2c4, apple2gs* and real apple2c0
- target `-t apple2enh` on emulated apple2ee, apple2ep, apple2c0, apple2c4, apple2gs* and real apple2c0

I couldn't test on apple2 and apple2p due to MAME having issues with the ROMs.
*(serial not tested on apple2gs as it's not the same chip)

**Numbers:**
Previous system had about:
- 248 cycles per ProDOS IRQ handler ($BFEB)
- 42 cycles per callirq call
- 62 cycles per ser_irq call

For a total of 352 cycles per ACIA interrupt (606 if triggered while in ROM, as $C803 has about 254 cycles before returning).

New system has:
- 37 cycles for handle_ram_irq
- no change in either callirq or ser_irq, as this isn't the scope of this patch

For a total of 141 cycles per ACIA interrupt (395 if triggered while in ROM).

**Callgrind profile before:**
![image](https://github.com/cc65/cc65/assets/1016317/0df7855c-7830-48fd-9705-560c3b00b2d4)

**Callgrind profile after:**
![image](https://github.com/cc65/cc65/assets/1016317/d380f3e9-8a37-4c9e-bc3e-1434335af260)

**Why:**
I'm on a quest to be able to use the serial port at 19200bps, and the existing IRQ handling is much too slow for this to succeed: another byte arrives before the IRQ for the previous one has returned. (This page shows how long it takes per byte to arrive: https://lucidar.me/en/serialib/most-used-baud-rates-table/)

**Results of this patch:**
With this patch, I am able to read/write in a sustained manner at 19200bps, **as long as RAM is banked in** when an IRQ arrives. If ROM is banked in, the added cost of the ROM IRQ handler, before execution is passed to the $03fe/$03ff vector, is still too much to sustain 19200 bps. But this means if one is cautious about how/when they receive bytes, double the speed can be achieved.

**Caveats:**
I am quite sure this patch is kind of naive, and there are probably corner cases that should be taken into account when entering the new IRQ handlers. I'd gladly improve the patch if it is necessary and it is deemed interesting to consider merging.

Thanks!