@colinleroy 2023-09-04
Let me add a comment about the different situations we can find ourselves in when an IRQ fires, because memory banking has been a concern on Apple2Infinitum's slack:

**Standard setup:**

"The vector at $FFFE in the ROM points to Apple lie’s built-in interrupt handling routine. When you initialize the mouse or serial communication firmware, copies of 
the ROM’s interrupt vector are placed in the interrupt vector addresses in both main and auxiliary bank-switched memory. [...]
"The built-in interrupt handling routine records the current memory configuration, then sets up its own standard memory configuration"
"passes them to a user’s interrupt handling routine whose address is stored at $03FE and $03FF of main memory".

On the Apple 2c for example, FFFE/FFFF point to $C803.

ProDOS then registers its handler into $3FE/3FF and it points to $BFEB.

so when an interrupt fires in the standard configuration, we have:
jmp ($fffe) => jmp $C803 => jmp ($3FE) => jmp $BFEB => callirq => our IRQ functions

and at the time callirq is called, we're setup to a known memory banking configuration that is triggered by ProDOS RR $c08b: read from RAM, write to RAM, LC bank 1. (it's the first thing it does, at $BFEB and $BFEE)

After this we go back to where we were, which in a cc65 program, is either c080 (read from RAM, write to none, LC bank 2) or c082 (read from ROM, write to none, LC bank 2).

**What this patches changes:**

1) it removes BFEB from 03FE and patches in handle_rom_irq instead, which only jsr to callirq.
2) it removes C803 from FFFE, __in the main RAM only__, and patches in handle_ram_irq instead, which saves registers, checks for BRK and jsr to callirq.

a) What happens when an IRQ fires and we're in ROM:
jmp ($fffe) => jmp $C803 => jmp ($3FE) => jmp handle_rom_irq => callirq => our IRQ functions
Is it bad that we're in ROM? No, because $C803 banks back to main RAM.

b) What happens when an IRQ fires and we're in main RAM:
jmp ($fffe) => jmp handle_ram_irq => callirq => our IRQ functions
Is it bad that we could be in LC1? No, because handle_rom_irq, callirq and our IRQ functions are all in LOWCODE or CODE, so always banked in.

c) What happens when an IRQ fires and we're in aux RAM:
jmp ($fffe) => jmp $C803 => jmp ($3FE) => jmp handle_rom_irq => callirq => our IRQ functions
Is it bad that we're in AUX RAM? No, because $C803 banks back to main RAM.

Finally, AppleIIgs always jumps to main RAM's ($fffe), so in case of a IIgs, we don't patch $fffe at all and we're always in case A or C. It's slower, but safe.