@colinleroy 2023-09-05
Closing this PR after discussion with @oliverschmidt : It is a bit too hacky to be deemed safe, and only applies to a small subset of Apple 2s.